requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.4
flake8==3.9.2
gunicorn==20.1.0
idna==3.2
mccabe==0.6.1
mongoengine==0.23.1
oauthlib==3.1.1
postgres==3.0.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
psycopg2-pool==1.1
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
pymongo==3.12.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
tweepy==3.10.0
urllib3==1.26.6
whitenoise==5.3.0
